// frontend/src/app/models/OffersModel.ts
export interface OffersModel {
    category?: string;
    categoryAr?: string;
    description?: string;
    descriptionAr?: string;
    merchant?: string;
    merchantAr?: string;
    endDate?: any;
    id?: number;
    primaryImage?: any;
    startDate?: any;
    termsAndConditions?: string;
    termsAndConditionsAr?: string;
    title?: string;
    titleAr?: string;
    websiteUrl?: string;
    websiteUrlAr?: string;
    images?: any[];
    location?: string;
    locationAr?: string;
    primaryImageFile?: any,
    imagesFiles?: any[];
}